+++ src/bin/documentation/readme-check.ts
@@ -4,6 +4,7 @@ import { githubReadmeContract } from "@/core/github/readme-contract";
 import { coreReadmeContract } from "@/core/readme-contract";
 import { checkReadmeFiles, handleReadmeCliError } from "@/features/docs";
 import { docsReadmeContract } from "@/features/docs/readme-contract";
+import { packagesReadmeContract } from "@/features/packages/readme-contract";
 import { releasesReadmeContract } from "@/features/releases/readme-contract";
 import { vitestReadmeContract } from "@/features/vitest/readme-contract";
 
@@ -13,10 +14,11 @@ await runTask(
     await checkReadmeFiles({
       "./README.md": rootReadmeContract,
       "./src/core/README.md": coreReadmeContract,
-      "./src/features/docs/README.md": docsReadmeContract,
       "./src/core/github/README.md": githubReadmeContract,
+      "./src/features/docs/README.md": docsReadmeContract,
       "./src/features/releases/README.md": releasesReadmeContract,
       "./src/features/vitest/README.md": vitestReadmeContract,
+      "./src/features/packages/README.md": packagesReadmeContract,
     }),
   handleReadmeCliError,
 );